Rename "database/types" folder to "database/models".
