+++ src/lib/database/manageRecipes.ts
@@ -1,7 +1,7 @@
 import type { Recipe, RecipeSubmit } from "$lib/domain/entities/Recipe";
 import { ENHANCEDCOFFEEBEANS_STORE, openEntitiesDB, RECIPES_COFFEEBEANSID_INDEX, RECIPES_STORE } from "./core/core";
 import { regenerateEnhancedCoffeeBeansItemById } from "./manageEnhancedCoffeeBeans";
-import { RecipeDB, type IRecipeDB, RecipeDBSubmit } from "./types/RecipeDB";
+import { RecipeDB, type IRecipeDB, RecipeDBSubmit } from "./models/RecipeDB";
 
 // Public functions:
