+++ src/lib/database/models/CoffeeBeansDB.ts
