+++ src/lib/database/models/EnhancedCoffeeBeansDB.ts
